Workaround asymmetry between distCache and distCache mocking
distCache requires that initFn is serializable, while distCache mock
requires that mocking value itself is serializable, which is not the
same. An example could be to a non-serializable client which requires
data to be available on all workers, pseudo code:

sc.distCache("init-data-file")(f => new NonSerializableClient(f))

Here NonSerializableClient does not need to be Serializable, but it's
not possible to test such job (at least without magic), even tho it's
perfectly fine job from distCache contract perspective.