Log exception in AsyncLookupDoFn if cache put fails (#1039)
If an exception is thrown when putting a value into the cache, the result will
not be added to the result queue, and finishBundle will fail the precondition
that requestCount == resultCount. The original exception happens in a transform
and does not get logged, so it's not clear to the user what went wrong.

This happens if the user caches a null value, which is not supported by the Guava
cache. This PR modifies the transform so the exception is logged.