Method to parse only expected options
I've found something like this useful when fixing a bug in
a project using scio. Basically, it protects you from issues
resulting from typos in options, or accidentally running an
older version that silently ignores an option.

When that came up, we solved it with a wrapper, but it would have
been nice if support for failing on unknown arguments was built
into scio.