Retry bulk chunk requests to Elasticsearch 2, 5, and 6
Implements retries when saving bulk documents to Elasticsearch.
This is useful for transient connection problems and other failures
which are temporary in nature.

The current behavior will cause the entire "Save to Elasticsearch" step
to fail and retry, which creates redundant saves to the cluster. There
is no existing method to enable a single failed chunk to retry.

This behavior is implemented with two new options, withMaxRetries and
withRetryPause, which can be passed in the configuration as "maxRetries"
and "retryPause" respectively.