Retry bulk chunk requests to Elasticsearch 2, 5, and 6
Implements retries when saving bulk documents to Elasticsearch.
This is useful for transient connection problems and other failures
which are temporary in nature.

The current behavior will cause the entire "Save to Elasticsearch" step
to fail and retry, which creates redundant saves to the cluster. There
is no existing method to enable a single failed chunk to retry.

This behavior is implemented with two new options, withMaxRetries and
withRetryPause, which can be passed in the configuration as "maxRetries"
and "retryPause" respectively.

Catch all exceptions within the chunks.forEach block.

Some of the exceptions thrown are not caught within the Client bulk() method.

Adds new params maxRetries and retryPause to saveAsElasticsearch method

Revert "Adds new params maxRetries and retryPause to saveAsElasticsearch method"

This exceeds the number of max params in scala style. So rely on it coming as cmdLine arg.

This reverts commit 00c299fbc2b5fbd823c697fc95e2bc0cc3f57cd9.

Implements FluentBackoff

Implements @Setup for ElasticsearchWriters, restores formatting

Restore indentation formatting on ES Bound constructors