Allow user to pass in Bigtable GcRule
Bigtable's GC policy can be configured to use the max age of a cell, the max
number of versions in a cell, or a combination of the two. Currently scio's
API only lets you specify max age. This PR makes it so the user can pass in
a GcRule, so they can configure GC however they like.