TestInput is not thread safe and can result in flaky tests.
TestInput's apply method calls add on a mutable Set as each
input is read. This call is not thread safe and can result in
some adds getting lost (probably when the underlying data structure
is resized). This results in flaky tests. This commit changes
the Sets used in TestDataManager to be thread safe.