Cache TensorFlow Metadata proto files
Previously, the build would download over and over the Tensorflow
metadata proto sources. On almost every build, close to a second
(depending, among others, on the connection speed) would  be spent
re-downloading and unzipping the same bytes. If the server where the
artifact is hosted became unavailable, developers would not be able to
build the project locally.

With this patch, this is no longer the case, as the proto files will be
cached locally. Cleaning the project, deleting or modifying a protobuf
source file will invalidate the cache and the sources will be downloaded
again.